requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
python-dotenv
requests